Add direction, circular and simplify to route segments
This allows route segments to be marked as circular and/or
bi-directional.

Circular means a route segment has no fixed start point: you can start
anywhere on the route segment.

Bi-directional means a route segment can be ran in both directions. This
is certainly not always the case.

The points along the route segment are also simplified before matching
against workouts; this means less points need to be checked, and the
chance is lower that the granularity of the workout is coarser than that
of the route segment.

The allowed total distance mismatch is now calculated as a percentage,
instead of an absolute distance.

Finally, we now show the tempo in the workout table.

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>